// prisma/generated/zod/inputTypeSchemas/ChatGPTHistoryUpdateInputSchema.ts
import type { Prisma } from '../..';

import { z } from 'zod';
import { StringFieldUpdateOperationsInputSchema } from './StringFieldUpdateOperationsInputSchema';
import { DateTimeFieldUpdateOperationsInputSchema } from './DateTimeFieldUpdateOperationsInputSchema';
import { UserUpdateOneRequiredWithoutChatHistoriesNestedInputSchema } from './UserUpdateOneRequiredWithoutChatHistoriesNestedInputSchema';

export const ChatGPTHistoryUpdateInputSchema: z.ZodType<Prisma.ChatGPTHistoryUpdateInput> = z.object({
  id: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  title: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  content: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  summary: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  createdAt: z.union([ z.coerce.date(),z.lazy(() => DateTimeFieldUpdateOperationsInputSchema) ]).optional(),
  updatedAt: z.union([ z.coerce.date(),z.lazy(() => DateTimeFieldUpdateOperationsInputSchema) ]).optional(),
  user: z.lazy(() => UserUpdateOneRequiredWithoutChatHistoriesNestedInputSchema).optional()
}).strict() as z.ZodType<Prisma.ChatGPTHistoryUpdateInput>;

export default ChatGPTHistoryUpdateInputSchema;